requirements.txt
MetaTrader5>=5.0.45   # Conexão com o MT5
pandas>=2.1.0         # Manipulação de dados (DataFrames)
numpy>=1.26.0         # Cálculos numéricos
pyyaml>=6.0           # Leitura do arquivo config.yaml
python-dotenv>=1.0.0  # Carregar credenciais do .env (Segurança)
# Para ML e Otimização
scikit-learn>=1.3.0
optuna>=3.4.0
matplotlib>=3.8.0
seaborn>=0.13.0